room3D

+++ src/genus/model_parts.py
@@ -307,7 +307,7 @@ class InferenceAndGeneration(torch.nn.Module):
                                                                     original_width=unet_prob_b1wh.shape[-2],
                                                                     original_height=unet_prob_b1wh.shape[-1])
 
-            score_grid = (1 - annealing_factor) * (c_grid_before_nms + unet_prob_b1wh) + \
+            score_grid = (1 - annealing_factor) * (c_grid_before_nms.float() + unet_prob_b1wh) + \
                          annealing_factor * prob_from_ranking_grid
             combined_topk_only = topk_only or generate_synthetic_data  # if generating from DPP do not do NMS
             nms_output: NmsOutput = NonMaxSuppression.compute_mask_and_index(score=convert_to_box_list(score_grid).squeeze(dim=-1),
